% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecCategorical.R
\name{codec_categorical}
\alias{codec_categorical}
\title{Estimate conditional dependence coefficient (CODEC) for categorical observation}
\usage{
codec_categorical(Y, Z, X = NULL, na.rm = TRUE, type = "continuous")
}
\arguments{
\item{Y}{Categorical or discrete value vector (length n)}

\item{Z}{Matrix (n by q)}

\item{X}{Matrix (n by p), default is NULL}

\item{na.rm}{Remove NAs if TRUE}

\item{type}{a factor variable indicating whether X and Z are continuous vectors or discrete, the
default value is continuous}
}
\value{
The conditional dependence coefficient (CODEC) of Y and Z given X. If X == NULL, this is just a
measure of the dependence between Y and Z.
}
\description{
The conditional dependence coefficient (CODEC) is a measure of the amount of conditional dependence between
a random variable Y and a random vector Z given a random vector X, based on an i.i.d. sample of (Y, Z, X).
The coefficient can take any real value in a finite sample, but is asymptotically guaranteed to be between 0 and 1.
}
\details{
The value returned by codec can be positive or negative. Asymptotically, it is guaranteed
to be between 0 and 1. A small value indicates low conditional dependence between Y and Z given X, and
a high value indicates strong conditional dependence. The codec_categorical function is used by the  \code{\link{foci_categorical}} function
for variable selection.
}
\examples{
n = 1000
p = 2
x = matrix(runif(n * p), ncol = p)
y = (2 * floor(2 * x[, 1]) - 1) * (2 * floor(2 * x[, 2]) - 1)
codec_categorical(y, x)
codec_categorical(y, x[, 1])
codec_categorical(y, x[, 2])
codec_categorical(y, x[, 1], x[, 2])
}
\references{
!!!!!!!!! NOT CLEAR WHERE YET :D !!!!!!!!!
}
\seealso{
\code{\link[FOCI]{codec}}, \code{\link[XICOR]{xicor}}
}
\author{
Mona Azadkia, Sourav Chatterjee, and Norman Matloff
}
