% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecInternal.R
\name{.estimateConditionalQ}
\alias{.estimateConditionalQ}
\title{.estimateConditionalQ -------------------------------------------------------------------------
Estimate Q(Y, Z | X)}
\usage{
.estimateConditionalQ(Y, X, Z)
}
\description{
.estimateConditionalQ -------------------------------------------------------------------------
Estimate Q(Y, Z | X)
}
