% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foci.R
\name{foci_main}
\alias{foci_main}
\title{Variable selection by the FOCI algorithm}
\usage{
foci_main(
  Y,
  X,
  num_features = NULL,
  stop = TRUE,
  numCores = parallel::detectCores()
)
}
\arguments{
\item{Y}{Vector of responses (length n)}

\item{X}{Matrix of predictors (n by p)}

\item{num_features}{Number of variables to be selected, cannot be larger than p. The default value is NULL and in that
case it will be set equal to p. If stop == TRUE (see below), then num_features is irrelevant.}

\item{stop}{Stops at the first instance of negative codec, if TRUE.}

\item{numCores}{number of cores that are going to be used for parallelizing the process.}
}
\value{
An object of class "foci", with three attributes a vector of selected covariates,
their names and their cumulative dependence value with Y in decreasing order of predictive power.
}
\description{
foci_main is the core function for foci, implementing the
variable selection algorithm based on the measure of conditional dependence \code{\link{codec}}.
}
\details{
foci is a forward stepwise algorithm that uses the conditional dependence coefficient (\code{\link{codec}})
at each step, instead of the multiple correlation coefficient
as in ordinary forward stepwise. If stop == TRUE, the process is stopped at the first instance of
nonpositive codec, thereby selecting subset of variables. Otherwise, a set of covariates of size
num_features, ordered according to predictive power (as measured by codec) is produced.
}
\references{
Azadkia, M. and Chatterjee, S. (2019). A simple measure of conditional dependence. \url{https://arxiv.org/pdf/1910.12327.pdf}
}
\seealso{
\code{\link{codec}}
}
\author{
Mona Azadkia, Sourav Chatterjee, and Norman Matloff
}
